pyproject.toml
[project]
name = "lead-processing-microservices"
version = "0.1.0"
description = "Two FastAPI microservices for lead ingestion and analytics."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.14.1",
    "asyncpg>=0.30.0",
    "fastapi>=0.115.8",
    "httpx>=0.28.1",
    "pydantic-settings>=2.8.0",
    "pyjwt>=2.10.1",
    "redis>=5.2.1",
    "sqlalchemy>=2.0.38",
    "uvicorn[standard]>=0.34.0",
]

[project.optional-dependencies]
dev = [
    "aiosqlite>=0.21.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["shared*", "landings*", "core*"]
